# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RU: API сервиса SVAP Query Service для интеграции со СВАП, управления сохраненными запросами, результатами и динамической конфигурацией. EN: API server for SVAP Query Service, used to integrate with SVAP and manage saved queries, query results, and dynamic configuration.",
        "title": "API SVAP Query Service / API агрегатора витрин",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/config": {
            "get": {
                "description": "RU: Возвращает всю подготовленную конфигурацию, включая еще не примененные изменения. EN: Returns the full pending configuration, including unapplied changes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "RU: Получить текущую подготовленную конфигурацию / EN: Get current pending configuration",
                "responses": {
                    "200": {
                        "description": "RU: Подготовленная конфигурация / EN: Pending configuration",
                        "schema": {
                            "$ref": "#/definitions/config.AppConfig"
                        }
                    }
                }
            },
            "put": {
                "description": "RU: Частично или полностью обновляет подготовленную конфигурацию из JSON-тела без немедленного применения. EN: Partially or fully updates the pending configuration from the JSON body without applying it immediately.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "RU: Обновить подготовленную конфигурацию / EN: Update pending configuration",
                "parameters": [
                    {
                        "description": "RU: Частичный или полный JSON конфигурации / EN: Partial or full configuration JSON",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {}
                    }
                ],
                "responses": {
                    "204": {
                        "description": "RU: Обновлено, тело ответа отсутствует / EN: Updated, no response body"
                    },
                    "400": {
                        "description": "RU: Некорректное тело запроса / EN: Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/config/apply": {
            "post": {
                "description": "RU: Сохраняет подготовленную конфигурацию в БД и отправляет уведомление узлам через PostgreSQL NOTIFY. EN: Persists the pending configuration to the database and notifies nodes through PostgreSQL NOTIFY.",
                "tags": [
                    "config"
                ],
                "summary": "RU: Применить подготовленную конфигурацию / EN: Apply pending configuration",
                "responses": {
                    "200": {
                        "description": "RU: Конфигурация применена / EN: Configuration applied"
                    },
                    "500": {
                        "description": "RU: Внутренняя ошибка сервера / EN: Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/config/groups/{name}": {
            "get": {
                "description": "RU: Возвращает настройки указанной группы: server, svap или retention. EN: Returns settings for the specified group: server, svap, or retention.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "config"
                ],
                "summary": "RU: Получить группу конфигурации / EN: Get a configuration group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RU: Имя группы / EN: Group name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RU: Настройки группы / EN: Group settings",
                        "schema": {}
                    },
                    "404": {
                        "description": "RU: Группа не найдена / EN: Group not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/queries": {
            "get": {
                "description": "RU: Возвращает список сохраненных запросов пользователя из заголовка X-User-ID. EN: Returns saved queries for the user from the X-User-ID header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queries"
                ],
                "summary": "RU: Получить сохраненные запросы / EN: List saved queries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RU: Идентификатор пользователя / EN: User identifier",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RU: Список сохраненных запросов / EN: Saved query list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.SavedQuery"
                            }
                        }
                    },
                    "401": {
                        "description": "RU: Не передан X-User-ID / EN: X-User-ID is missing",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "RU: Внутренняя ошибка сервера / EN: Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "RU: Регистрирует новый шаблон запроса с параметрами фильтрации. Пользователь определяется по заголовку X-User-ID. EN: Registers a new query template with filter parameters. The user is resolved from the X-User-ID header.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "queries"
                ],
                "summary": "RU: Создать сохраненный запрос / EN: Create a saved query",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RU: Идентификатор пользователя / EN: User identifier",
                        "name": "X-User-ID",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "RU: Описание запроса / EN: Query definition",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SavedQuery"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "RU: Запрос создан / EN: Query created",
                        "schema": {
                            "$ref": "#/definitions/model.SavedQuery"
                        }
                    },
                    "400": {
                        "description": "RU: Некорректное тело запроса или видимость / EN: Invalid request body or visibility",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "RU: Не передан X-User-ID / EN: X-User-ID is missing",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "RU: user_id не совпадает с X-User-ID / EN: user_id does not match X-User-ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "RU: Внутренняя ошибка сервера / EN: Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/queries/{id}/results": {
            "get": {
                "description": "RU: Возвращает историю выполнений для указанного сохраненного запроса. EN: Returns execution history for the specified saved query.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "results"
                ],
                "summary": "RU: Получить результаты запроса / EN: List query results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RU: UUID запроса / EN: Query UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RU: История результатов / EN: Query result history",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.QueryResult"
                            }
                        }
                    },
                    "400": {
                        "description": "RU: Некорректный UUID запроса / EN: Invalid query UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "RU: Внутренняя ошибка сервера / EN: Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/results/{id}": {
            "delete": {
                "description": "RU: Удаляет конкретный результат выполнения запроса и связанные строки данных. EN: Deletes a specific query execution result and its stored data rows.",
                "tags": [
                    "results"
                ],
                "summary": "RU: Удалить результат запроса / EN: Delete a query result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "RU: UUID результата / EN: Result UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "RU: Удалено, тело ответа отсутствует / EN: Deleted, no response body"
                    },
                    "400": {
                        "description": "RU: Некорректный UUID результата / EN: Invalid result UUID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "RU: Внутренняя ошибка сервера / EN: Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "config.AppConfig": {
            "type": "object",
            "properties": {
                "retention": {
                    "$ref": "#/definitions/config.RetentionConfig"
                },
                "server": {
                    "$ref": "#/definitions/config.ServerConfig"
                },
                "svap": {
                    "$ref": "#/definitions/config.SVAPConfig"
                }
            }
        },
...
        "config.ServerConfig": {
            "type": "object",
            "properties": {
                "port": {
                    "type": "string"
                }
            }
        },
...
        "model.SavedQuery": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "params": {
                    "description": "JSON-строка параметров",
                    "type": "string"
                },
                "query_type": {
                    "description": "OPLIST, FSG, TURN, ReportGK, COR, PA, CONS",
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "visibility": {
                    "description": "private, organization, public",
                    "type": "string"
                }
            }
        }
    }
}